[Core] Split frozen<T> and message_queue<T>.
Signals don't use queues, but use frozen events. Split these two entities to
reduce include burden.
