+++ include/cocaine/rpc/dispatch.hpp
@@ -105,7 +105,9 @@ class dispatch:
 
     typedef typename mpl::transform<
         typename io::messages<Tag>::type,
-        typename mpl::lambda<std::shared_ptr<io::basic_slot<mpl::_1>>>::type
+        typename mpl::lambda<
+            std::shared_ptr<io::basic_slot<mpl::_1>>
+        >::type
     >::type slot_types;
 
     typedef std::map<
